+++ src/_load_log.jl
@@ -25,6 +25,7 @@ function load_log(filename::String; path="")
     end
     table   = Arrow.Table(fullname)
     P =  length(table.Z[1])
+    O =  length(table.Qw[1])
     colmeta = Dict(:var_01=>Arrow.getmetadata(table.var_01)["name"],
                    :var_02=>Arrow.getmetadata(table.var_02)["name"],
                    :var_03=>Arrow.getmetadata(table.var_03)["name"],
@@ -43,19 +44,19 @@ function load_log(filename::String; path="")
                    :var_16=>Arrow.getmetadata(table.var_16)["name"],
     )
     # example_metadata = KiteUtils.Arrow.getmetadata(table.var_01)
-    syslog = StructArray{SysState{P}}((table.time, table.t_sim, table.sys_state, table.cycle, table.fig_8, 
-                                       table.e_mech, table.orient, table.turn_rates, table.elevation, table.azimuth, 
-                                       table.azimuth_rate, table.l_tether, table.v_reelout, table.winch_force, table.depower, 
-                                       table.steering, table.kcu_steering, table.set_steering, table.heading, table.heading_rate, 
-                                       table.course, table.bearing, table.attractor, table.v_app, table.v_wind_gnd, 
-                                       table.v_wind_200m, table.v_wind_kite, table.AoA, table.side_slip, table.alpha3, 
-                                       table.alpha4, table.CL2, table.CD2, table.aero_force_b, table.aero_moment_b, 
-                                       table.tether_induced_force, table.tether_induced_moment, table.twist_angles, table.vel_kite, table.acc, 
-                                       table.X, table.Y, table.Z, table.set_torque, table.set_speed, 
-                                       table.set_force, table.roll, table.pitch, table.yaw, table.var_01, 
-                                       table.var_02, table.var_03, table.var_04, table.var_05, table.var_06, 
-                                       table.var_07, table.var_08, table.var_09, table.var_10, table.var_11, 
-                                       table.var_12, table.var_13, table.var_14, table.var_15, table.var_16
-                                       ))
+    syslog = StructArray{SysState{P, O}}((table.time, table.t_sim, table.sys_state, table.cycle, table.fig_8, 
+                                       table.e_mech, table.Qw, table.Qx, table.Qy, table.Qz, 
+                                       table.turn_rates, table.elevation, table.azimuth, table.azimuth_rate, table.l_tether, 
+                                       table.v_reelout, table.winch_force, table.depower, table.steering, table.kcu_steering, 
+                                       table.set_steering, table.heading, table.heading_rate, table.course, table.bearing, 
+                                       table.attractor, table.v_app, table.v_wind_gnd, table.v_wind_200m, table.v_wind_kite, 
+                                       table.AoA, table.side_slip, table.alpha3, table.alpha4, table.CL2, 
+                                       table.CD2, table.aero_force_b, table.aero_moment_b, table.tether_induced_force, table.tether_induced_moment, 
+                                       table.twist_angles, table.vel_kite, table.acc, table.X, table.Y, 
+                                       table.Z, table.set_torque, table.set_speed, table.set_force, table.roll, 
+                                       table.pitch, table.yaw, table.var_01, table.var_02, table.var_03, 
+                                       table.var_04, table.var_05, table.var_06, table.var_07, table.var_08, 
+                                       table.var_09, table.var_10, table.var_11, table.var_12, table.var_13, 
+                                       table.var_14, table.var_15, table.var_16))
     return SysLog{P}(basename(fullname[1:end-6]), colmeta, syslog)
 end